+++ examples/serverless/src/dispatcher.js
@@ -1,8 +1,8 @@
 
-exports.handler = async (event, context) => {
+exports.handler = (event, context, callback) => {
     context.callbackWaitForEmptyEventLoop = false;
-    console.log(event.Records[0].Sns)
-    return {
+    console.log(event.Records[0].Sns);
+    return callback(null, {
       message: 'req processed'
-    }
+    })
 };