+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.18] 2023-07-24
+
+- Improved robustness of callback support
 
 ## [0.1.17] 2023-07-24